+++ src/v2019C1/extensions/data-model/public.ts
@@ -1,7 +1,5 @@
-export { defaultForkId } from './transaction/import-types'
 export type {
 	ImportTypesParams,
 	ImportTypesResult,
 	ImportTypesStats,
-	ForkIdContext,
 } from './transaction/import-types'